+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 0.12
 - changed dl_write_access 'ADMIN TOKEN' option to 'TOKEN', meaning ADMIN REQUIRING TOKEN FOR UPDATE
 - renames: domain datalink.dl_file_path is now datalink.file_path
 - improved checks on domain datalink.file_path
+- alter table rename column fixes
 
 Version 0.11
 ------------